+++ backend/pkg/git/service.go
@@ -3,6 +3,7 @@ package git
 import (
 	"context"
 	"fmt"
+	"github.com/cloudhut/kowl/backend/pkg/filesystem"
 	"github.com/go-git/go-billy/v5"
 	"github.com/go-git/go-billy/v5/memfs"
 	"github.com/go-git/go-git/v5"
@@ -30,20 +31,10 @@ type Service struct {
 	repo  *git.Repository
 
 	// In memory cache for markdowns. Map key is the filename with stripped ".md" suffix.
-	filesByName map[string]File
+	filesByName map[string]filesystem.File
 	mutex       sync.RWMutex
 
 	OnFilesUpdatedHook func()
-}
-
-type File struct {
-	Path     string
-	Filename string
-
-	// TrimmedFilename is the filename without the recognized file extension
-	TrimmedFilename string
-
-	Payload []byte
 }
 
 // NewService creates a new Git service with preconfigured Auth
@@ -72,7 +63,7 @@ func NewService(cfg Config, logger *zap.Logger, onFilesUpdatedHook func()) (*Ser
 		auth:   auth,
 		logger: childLogger,
 
-		filesByName:        make(map[string]File),
+		filesByName:        make(map[string]filesystem.File),
 		OnFilesUpdatedHook: onFilesUpdatedHook,
 	}, nil
 }
@@ -116,7 +107,7 @@ func (c *Service) CloneRepository(ctx context.Context) error {
 	c.repo = repo
 
 	// 2. Put files into cache
-	empty := make(map[string]File)
+	empty := make(map[string]filesystem.File)
 	files, err := c.readFiles(fs, empty, ".", 5)
 	if err != nil {
 		return fmt.Errorf("failed to get files: %w", err)
@@ -168,7 +159,7 @@ func (c *Service) SyncRepo() {
 			}
 
 			// Update cache with new markdowns
-			empty := make(map[string]File)
+			empty := make(map[string]filesystem.File)
 			files, err := c.readFiles(c.memFs, empty, ".", 5)
 			if err != nil {
 				c.logger.Error("failed to read files after pulling", zap.Error(err))
@@ -187,7 +178,7 @@ func (c *Service) SyncRepo() {
 
 // setFileContents saves file contents into memory, so that they are accessible at any time.
 // filesByName is a map where the key is the filename without extension suffix (e.g. "README" instead of "README.md")
-func (c *Service) setFileContents(filesByName map[string]File) {
+func (c *Service) setFileContents(filesByName map[string]filesystem.File) {
 	c.mutex.Lock()
 	defer c.mutex.Unlock()
 
@@ -197,20 +188,20 @@ func (c *Service) setFileContents(filesByName map[string]File) {
 // GetFileByFilename returns the cached content for a given filename (without extension).
 // The parameter must match the filename in the git repository (case sensitive).
 // If there's no match nil will be returned.
-func (c *Service) GetFileByFilename(fileName string) File {
+func (c *Service) GetFileByFilename(fileName string) filesystem.File {
 	c.mutex.RLock()
 	defer c.mutex.RUnlock()
 
 	contents, exists := c.filesByName[fileName]
 	if !exists {
-		return File{}
+		return filesystem.File{}
 	}
 
 	return contents
 }
 
 // GetFilesByFilename returns the cached content in a map where the filename is the key (with trimmed file extension).
-func (c *Service) GetFilesByFilename() map[string]File {
+func (c *Service) GetFilesByFilename() map[string]filesystem.File {
 	c.mutex.RLock()
 	defer c.mutex.RUnlock()