+++ backend/pkg/schema/config_tls.go
@@ -1,6 +1,10 @@
 package schema
 
-// TLSConfig to connect to Kafka via TLS
+// TLSConfig to connect to Schema via TLS
 type TLSConfig struct {
-	CaFilepath string `yaml:"caFilepath"`
+	Enabled               bool   `yaml:"enabled"`
+	CaFilepath            string `yaml:"caFilepath"`
+	CertFilepath          string `yaml:"certFilepath"`
+	KeyFilepath           string `yaml:"keyFilepath"`
+	InsecureSkipTLSVerify bool   `yaml:"insecureSkipTlsVerify"`
 }