+++ frontend/src/components/misc/Card.tsx
@@ -1,10 +1,10 @@
 import React, { ReactNode, Component, CSSProperties } from "react";
 
 
-class Card extends Component<{ style?: CSSProperties, className?: string }> {
+class Card extends Component<{ id?: string, style?: CSSProperties, className?: string }> {
 
     render() {
-        return <div className={'kowlCard ' + (this.props.className ?? '')} style={this.props.style}>
+        return <div id={this.props.id} className={'kowlCard ' + (this.props.className ?? '')} style={this.props.style}>
             {this.props.children}
         </div>
     }