+++ frontend/src/utils/numberExtensions.ts
@@ -9,10 +9,18 @@ declare global {
          * @param t factor between 0 and 1
          */
         lerp(this: number, to: number, t: number): number;
+
+        clamp(this: number, min: number, max: number): number;
     }
 }
 
 Number.prototype.lerp = function (this: number, to: number, t: number): number {
     const d = to - this;
     return this + d * t;
+};
+
+Number.prototype.clamp = function (this: number, min: number, max: number): number {
+    if (this > max) return max;
+    if (this < min) return min;
+    return this;
 };